EWPP-1822: Add some code corrections and reorganise code.
